revised fileopd

+++ lib/fileop/impl/standalone.py
@@ -28,7 +28,6 @@ class StandaloneFileOperation(FileTransferOperation, FileTransferQuery, FileDele
 
         if hasattr(tasks[0], 'source'):
             # These are transfer tasks
-            # GFAL2 has a "bulk-copy" feature which supposedly is more efficient than copying one by one
             by_endpoints = collections.defaultdict(list)
             for task in tasks:
                 endpoints = (task.source, task.subscription.destination)
@@ -43,16 +42,25 @@ class StandaloneFileOperation(FileTransferOperation, FileTransferQuery, FileDele
             return by_endpoint.values()
 
     def start_transfers(self, batch_id, batch_tasks): #override
+        if len(batch_tasks) == 0:
+            return True
+
+        # tasks should all have the same source and destination
+        source = batch_tasks[0].source
+        destination = batch_tasks[0].subscription.destination
+
         fields = ('id', 'source', 'destination')
         def mapping(task):
             lfn = task.subscription.file.lfn
             return (
                 task.id,
-                task.source.to_pfn(lfn, 'gfal2'),
-                task.subscription.destination.to_pfn(lfn, 'gfal2')
+                source.to_pfn(lfn, 'gfal2'),
+                destination.to_pfn(lfn, 'gfal2')
             )
 
         if not self.dry_run:
+            sql = 'INSERT INTO `standalone_transfer_batches` (`batch_id`, `source_site`, `destination_site`) VALUES (%s, %s, %s)'
+            self.db.query(sql, batch_id, source.name, destination.name)
             self.db.insert_many('standalone_transfer_queue', fields, mapping, batch_tasks)
 
         LOG.debug('Inserted %d entries to standalone_transfer_queue for batch %d.', len(batch_tasks), batch_id)
@@ -60,15 +68,23 @@ class StandaloneFileOperation(FileTransferOperation, FileTransferQuery, FileDele
         return True
 
     def start_deletions(self, batch_id, batch_tasks): #override
+        if len(batch_tasks) == 0:
+            return True
+
+        # tasks should all have the same target site
+        site = batch_tasks[0].desubscription.site
+
         fields = ('id', 'file')
         def mapping(task):
-            lfn = task.subscription.file.lfn
+            lfn = task.desubscription.file.lfn
             return (
                 task.id,
-                task.desubscription.site.to_pfn(lfn, 'gfal2')
+                site.to_pfn(lfn, 'gfal2')
             )
 
         if not self.dry_run:
+            sql = 'INSERT INTO `standalone_deletion_batches` (`batch_id`, `site`) VALUES (%s, %s)'
+            self.db.query(sql, batch_id, site.name)
             self.db.insert_many('standalone_deletion_queue', fields, mapping, batch_tasks)
 
         LOG.debug('Inserted %d entries to standalone_deletion_queue for batch %d.', len(batch_tasks), batch_id)
@@ -96,6 +112,16 @@ class StandaloneFileOperation(FileTransferOperation, FileTransferQuery, FileDele
         return [(i, FileQuery.status_val(s), c, t, f) for (i, s, c, t, f) in self.db.xquery(sql, batch_id)]
 
     def _forget_status(self, batch_id, task_id, optype):
+        if self.dry_run:
+            return
+
         sql = 'DELETE FROM `standalone_{op}_queue` WHERE `id` = %s'
         sql = sql.format(op = optype)
         self.db.query(sql, task_id)
+
+        sql = 'SELECT COUNT(*) FROM `standalone_{op}_queue` AS a'
+        sql += ' INNER JOIN `{op}_queue` AS q ON q.`id` = a.`id`'
+        sql += ' WHERE q.`batch_id` = %s'
+        if self.db.query(sql.format(op = optype), batch_id)[0] == 0:
+            sql = 'DELETE FROM `standalone_{op}_batches WHERE `batch_id` = %s`'
+            self.db.query(sql.format(op = optype), batch_id)
